backend/requirements.txt
fastapi
uvicorn
selenium
pydantic
typing
SQLAlchemy
psycopg2
spacy
PyPDF2  
python-multipart
python-dotenv
langchain
langchain-community
langchain-groq
langchain-core
langchain-google-genai
pgvector
sentence-transformers
faiss-cpu
numpy
langchain-huggingface
selectorlib
lxml
parsel
requests
playwright
nlp
chromium
spacy
datasets


